+++ guillotina/test_package.py
@@ -15,6 +15,7 @@ from guillotina.directives import write_permission
 from guillotina.exceptions import NoIndexField
 from guillotina.fields import CloudFileField
 from guillotina.files import BaseCloudFile
+from guillotina.files.exceptions import RangeNotFound
 from guillotina.files.utils import generate_key
 from guillotina.interfaces import IApplication
 from guillotina.interfaces import IContainer
@@ -31,6 +32,7 @@ from guillotina.response import HTTPUnprocessableEntity
 from guillotina.schema import Object
 from guillotina.schema.interfaces import IContextAwareDefaultFactory
 from shutil import copyfile
+from typing import AsyncIterator
 from zope.interface import implementer
 from zope.interface import Interface
 
@@ -320,11 +322,27 @@ class InMemoryFileManager:
         _tmp_files[upload_file_id] = tempfile.mkstemp()[1]
         await dm.update(_chunks=0, upload_file_id=upload_file_id)
 
-    async def delete_upload(self, uri):
+    async def delete_upload(self, uri):  # pragma: no cover
         if uri in _tmp_files:
             if os.path.exists(_tmp_files[uri]):
                 os.remove(_tmp_files[uri])
             del _tmp_files[uri]
+
+    async def range_supported(self) -> bool:
+        return True
+
+    async def read_range(self, start: int, end: int) -> AsyncIterator[bytes]:
+        file = self.field.get(self.field.context or self.context)
+        uri = file.uri
+        total = 0
+        with open(_tmp_files[uri], "rb") as fi:
+            fi.seek(start)
+            while total < (end - start):
+                chunk = fi.read(1024 * 1024)
+                total += len(chunk)
+                yield chunk
+        if len(chunk) != (end - start):
+            raise RangeNotFound(field=self.field, start=start, end=end)
 
     async def append(self, dm, iterable, offset) -> int:
         count = 0