+++ CHANGELOG.rst
@@ -4,12 +4,14 @@ CHANGELOG
 6.0.0a5 (unreleased)
 --------------------
 
+- Add `Range` header support
+  [vangheem]
+
 - Fix validating None values in required fields
   [vangheem]
 
 - Add localroles to @available-roles
   [jordic]
-
 
 6.0.0a4 (2019-12-23)
 --------------------